+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM
+### Fixed
+- Fix survey display issues [#5](https://github.com/rokmetro/neom-app/issues/5)
 
 ## [0.4.0] - 2024-07-26
 ### Fixed